+++ Source/RequestBuilder.swift
@@ -20,16 +20,13 @@ import Foundation
 /**
  Designates an operation which provides data to perform a HTTP Request.
  */
-protocol HTTPRequestOperation {
+internal protocol HTTPRequestOperation   {
 
-    /**
-     Provides the `InterceptableSession` to use when making HTTP requests.
-     */
-    var session: InterceptableSession { get }
     /**
      The root of url, e.g. `example.cloudant.com`
      */
     var rootURL: NSURL { get }
+    
     /**
      The path of the url e.g. `/exampledb/document1/`
      */
@@ -42,18 +39,23 @@ protocol HTTPRequestOperation {
      The query items to use for the request
      */
     var queryItems: [NSURLQueryItem] { get }
-
+    
     /**
      The body of the HTTP request or `nil` if there is no data for the request.
      */
     var httpRequestBody: NSData? { get }
-
+    
     /**
      The content type of the HTTP request payload. This is guranteed to be called
      if and only if `httpRequestBody` is not `nil`
      */
     var httpContentType: String { get }
-
+    
+    /**
+     Provides the `InterceptableSession` to use when making HTTP requests.
+     */
+    var session: InterceptableSession { get }
+    
     /**
      A function that is called when the operation is completed.
      */